[update] using kTol=1e-4 and GpuT=float by default

+++ multibody/gpu_mpm/settings.h
@@ -34,7 +34,7 @@ template<typename T> using Vec2 = Eigen::Vector<T, 2>;
 template<typename T> using Mat2 = Eigen::Matrix<T, 2, 2>;
 
 namespace config {
-	using GpuT = double;
+	using GpuT = float;
 
     // cuda device
     constexpr int G_DEVICE_COUNT = 1;
